Creating function result price

+++ src/main/kotlin/com/example/testtask/service/IShopService.kt
@@ -5,7 +5,7 @@ import com.example.testtask.controller.PurchaseDto
 
 interface IShopService {
 
-    fun calculatePrice(dto: CalculatePriceDto):UInt
+    fun calculatePrice(dto: CalculatePriceDto): Double
 
     fun purchase(dto: PurchaseDto):String
